Renaming the Navbar folder
