+++ compounds/HeaderCompound.js
@@ -1,4 +1,4 @@
-import HeaderContainer from '../components/Navbar/HeaderContainer'
+import HeaderContainer from '../components/Header/HeaderContainer'
 
 export default function HeaderCompound() {
   return (
